requirements.txt
apache-airflow==2.7.1
numpy==1.24.2
pandas==2.0.0
psycopg2==2.9.7
Requests==2.31.0
SQLAlchemy==2.0.20
